pyproject.toml
[build-system]  # https://docs.astral.sh/uv/concepts/projects/config/#build-systems
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]  # https://packaging.python.org/en/latest/specifications/pyproject-toml/
name = "minifigures-app"
version = "0.0.0"
description = "The goal of this project is to teach the students everything they need to develop and deploy an AI solution from start to finish."
readme = "README.md"
authors = [
  { name = "Adriaan Van Haecke", email = "adriaan@superlinear.eu" },
]
requires-python = ">=3.10,<3.11"
dependencies = [
  "poethepoet (>=0.32.1)",
  "dvc (>=2.40.0)",
  "dvc-s3 (>=2.21.0)",
  "fastapi (>=0.78.0)",
  "gunicorn (>=20.1.0)",
  "llvmlite (>=0.41.1)",
  "numba (>=0.58.0)",
  "onnxruntime (>=1.13.1)",
  "opencv-python-headless (>=4.7.0.68)",
  "pooch (>=1.6.0)",
  "pymatting (>=1.1.8)",
  "python-multipart (>=0.0.5)",
  "streamlit (>=1.16.0)",
  "uvicorn[standard] (>=0.17.0)",
  "scikit-multilearn (>=0.2.0)",
  "torch @ https://download.pytorch.org/whl/cpu/torch-1.13.1%2Bcpu-cp310-cp310-linux_x86_64.whl",
  "torchvision @ https://download.pytorch.org/whl/cpu/torchvision-0.14.1%2Bcpu-cp310-cp310-linux_x86_64.whl",
  "pillow>=12.0.0",
]

[project.urls]  # https://packaging.python.org/en/latest/specifications/well-known-project-urls/#well-known-labels
homepage = "https://github.com/kuleuven-realization-of-ai/minifigures-webshop-template-uv"
source = "https://github.com/kuleuven-realization-of-ai/minifigures-webshop-template-uv"
changelog = "https://github.com/kuleuven-realization-of-ai/minifigures-webshop-template-uv/blob/main/CHANGELOG.md"
releasenotes = "https://github.com/kuleuven-realization-of-ai/minifigures-webshop-template-uv/releases"
documentation = "https://kuleuven-realization-of-ai.github.io/minifigures-webshop-template-uv"
issues = "https://github.com/kuleuven-realization-of-ai/minifigures-webshop-template-uv/issues"

[dependency-groups]  # https://docs.astral.sh/uv/concepts/projects/dependencies/#development-dependencies
dev = [
  "commitizen (>=4.3.0)",
  "coverage[toml] (>=7.6.10)",
  "ipykernel (>=6.29.4)",
  "ipython (>=8.18.0)",
  "ipywidgets (>=8.1.2)",
  "mypy (>=1.14.1)",
  "pre-commit (>=4.0.1)",
  "pytest (>=8.3.4)",
  "pytest-mock (>=3.14.0)",
  "pytest-xdist (>=3.6.1)",
  "ruff (>=0.9.2)",
]
train = [
  "albumentations (>=1.3.0)",
  "matplotlib (>=3.6.2)",
  "tensorboard (>=2.11.0)",
  "label-studio (==1.11.0)",
  "label-studio-sdk (>=0.0.32)",
  "tqdm>=4.67.1",
  "fastdup>=2.47",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.commitizen]  # https://commitizen-tools.github.io/commitizen/config/
bump_message = "bump: v$current_version → v$new_version"
tag_format = "v$version"
update_changelog_on_bump = true
version_provider = "uv"

[tool.coverage.report]  # https://coverage.readthedocs.io/en/latest/config.html#report
precision = 1
show_missing = true
skip_covered = true

[tool.coverage.run]  # https://coverage.readthedocs.io/en/latest/config.html#run
branch = true
command_line = "--module pytest"
data_file = "reports/.coverage"
source = ["src"]

[tool.coverage.xml]  # https://coverage.readthedocs.io/en/latest/config.html#xml
output = "reports/coverage.xml"

[tool.mypy]  # https://mypy.readthedocs.io/en/latest/config_file.html
junit_xml = "reports/mypy.xml"
ignore_missing_imports = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
warn_unreachable = true

[tool.pytest.ini_options]  # https://docs.pytest.org/en/latest/reference/reference.html#ini-options-ref
addopts = "--color=yes --doctest-modules --exitfirst --failed-first --verbosity=2 --junitxml=reports/pytest.xml"
testpaths = ["src", "tests"]
xfail_strict = true

[tool.ruff]  # https://docs.astral.sh/ruff/settings/
fix = true
line-length = 100
src = ["src", "tests"]
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true
skip-magic-trailing-comma = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["CPY", "FIX", "T20", "ARG001", "COM812", "D203", "D213", "E501", "PD008", "PD009", "PGH003", "RET504", "S101", "TD003"]
unfixable = ["ERA001", "F401", "F841", "T201", "T203"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true
ignore-fully-untyped = true

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.pycodestyle]
max-doc-length = 100

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.poe.executor]  # https://github.com/nat-n/poethepoet
type = "simple"

[tool.poe.tasks]

  [tool.poe.tasks.api]
  help = "Run the REST API"
  shell = """
    if [ $dev ]
    then {
      uvicorn \
        --host $host \
        --port $port \
        --reload \
        minifigures_api.api:app
    } else {
    gunicorn \
      --access-logfile - \
      --bind $host:$port \
      --graceful-timeout 10 \
      --keep-alive 10 \
      --log-file - \
      --timeout 30 \
      --worker-class uvicorn.workers.UvicornWorker \
      --worker-tmp-dir /dev/shm \
      --workers 2 \
      minifigures_api.api:app
    } fi
    """

    [[tool.poe.tasks.api.args]]
    help = "Bind socket to this host (default: 0.0.0.0)"
    name = "host"
    options = ["--host"]
    default = "0.0.0.0"

    [[tool.poe.tasks.api.args]]
    help = "Bind socket to this port (default: 8000)"
    name = "port"
    options = ["--port"]
    default = "8000"

    [[tool.poe.tasks.api.args]]
    help = "Run in development mode"
    type = "boolean"
    name = "dev"
    options = ["--dev"]

  [tool.poe.tasks.app]
  help = "Serve the Streamlit application"
  cmd = """
    streamlit run
      --browser.gatherUsageStats false
      --server.address $host
      --server.port $port
      --theme.base light
      --server.enableCORS false
      src/minifigures_app/🏠_Home.py
    """
  use_exec = true

    [[tool.poe.tasks.app.args]]
    help = "Bind socket to this host (default: 0.0.0.0)"
    name = "host"
    options = ["--host"]
    default = "0.0.0.0"

    [[tool.poe.tasks.app.args]]
    help = "Bind socket to this port (default: 8500)"
    name = "port"
    options = ["--port"]
    default = "8500"

  [tool.poe.tasks.lint]
  help = "Lint this app"
  cmd = """
    pre-commit run
      --all-files
      --color always
    """

  [tool.poe.tasks.test]
  help = "Test this app"

    [[tool.poe.tasks.test.sequence]]
    cmd = "coverage run"

    [[tool.poe.tasks.test.sequence]]
    cmd = "coverage report"

    [[tool.poe.tasks.test.sequence]]
    cmd = "coverage xml"

[tool.uv]
# This tells uv to only worry about Linux (fixes the 'darwin' errors)
environments = ["sys_platform == 'linux'"]

# This forces fastdup to accept the versions the rest of your project needs
override-dependencies = [
    "requests>=2.31.0",
    "tqdm>=4.67.1",
    "numpy>=1.24.3",
    "pillow>=10.3.0",
    "opencv-python-headless==4.8.0.74",
]
